refactor: create app/server directory
